+++ NEWS.md
@@ -1,3 +1,7 @@
+## CHANGES IN audio.vadwebrtc VERSION 0.2
+
+- Added function VAD_channel to detect voice in audio by channel which depends on reading/writing audio data with R package audio and converting an audio file to a specific sample_rate with R package av
+
 ## CHANGES IN audio.vadwebrtc VERSION 0.1
 
 - Added function VAD to detect voice in audio